Move lib/launcher.rb to lib/proxy/launcher.rb
This matches the module layout
