+++ bin/smart-proxy
@@ -3,4 +3,5 @@
 $LOAD_PATH.unshift(*Dir[File.expand_path('../lib', __dir__), File.expand_path('../modules', __dir__)])
 
 require 'smart_proxy_main'
+require 'proxy/launcher'
 Proxy::Launcher.new.launch
